functions/requirements.txt
requests==2.22.0
minio==5.0.4
Pillow==7.1.0
boto3==1.10.9
azure==4.0.0
azure-storage==0.36.0
azure-functions==1.0.7
onnxruntime==1.1.0
numpy==1.18.1
python-resize-image
opencv-python
imutils